requirements.txt
fastapi==0.111.0
uvicorn==0.30.0

openai==1.35.0
google-generativeai==0.7.2
anthropic==0.34.0
groq==0.9.0
httpx==0.27.2

pymongo==4.7.2
dnspython==2.6.1

wikipedia==1.4.0
requests==2.32.3
duckduckgo-search==8.1.1

streamlit==1.36.0

python-dotenv==1.0.1
